requirements.txt
fastapi[all]
masonite-orm
passlib
bcrypt[cryptography]
psycopg2-binary
fastapi-mail
python-dotenv
pytest
requests